Display all required but missing options in help message
If more than one parameter was specified as required but missing, only
the first in the "Option" class was displayed as in error. Now
displaying all of them.

Also changed reflection utility to get either the entry assembly
(current) or the executing assembly. This was throwing exceptions
running tests from the CodeRush test runner.