Implemented Hidden Verbs, Options and Values
The attribute Hidden has been added to OptionAttribute, ValueAttribute,
and VerbAttribute.  By default it is false.

Changes to support this new attribute so that if hidden it is not
visible in the help text, but is still usable if called correctly. If
you ask for help on a hidden verb you will get the help displayed but
hidden options will still be hidden. Allowing for hidden verbs with not
hidden options.

Resolves issue #263